added discription
Undead creatures are attacking the city! Position your units to protect the city until you have killed enough of the creatures to drive them off.

In this “tower defence” strategy game, monsters are advancing on the city from right to left across 5 lanes. To kill the monsters, you have to purchase units and place them on the field of battle so that they can shoot or block the monsters. However, you start with 10 gold and only get 1 gold per turn, so spend your precious resources wisely!